dockershim: move naming helpers to a separate file
